Fixed parsing for the number literals to report more errors
The code for number parsing was redesigned to report more error,
also it is now uses standard TextPos class to report position of error.
Also fixed LexerImpl to use common character utilities with LiteralParser.

+++ etl-parser/src/main/java/net/sf/etl/parsers/TextPos.java
@@ -1,26 +1,26 @@
 /*
  * Reference ETL Parser for Java
- * Copyright (c) 2000-2009 Constantine A Plotnikov
+ * Copyright (c) 2000-2012 Constantine A Plotnikov
  *
- * Permission is hereby granted, free of charge, to any person 
- * obtaining a copy of this software and associated documentation 
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
  * files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, 
- * publish, distribute, sublicense, and/or sell copies of the Software, 
- * and to permit persons to whom the Software is furnished to do so, 
+ * including without limitation the rights to use, copy, modify, merge,
+ * publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so,
  * subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be 
+ * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, 
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF 
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND 
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
  * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN 
- * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE 
- * SOFTWARE. 
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 package net.sf.etl.parsers;
 
@@ -55,13 +55,13 @@ public final class TextPos implements Serializable {
      */
     private final int column;
     /**
-     * offset
+     * offset in characters
      */
     private final long offset;
     /**
      * start offset
      */
-    public static final int START_OFFSET = 0;
+    public static final long START_OFFSET = 0;
     /**
      * start column
      */
@@ -90,7 +90,7 @@ public final class TextPos implements Serializable {
     }
 
     /**
-     * @return column
+     * @return column in symbols
      */
     public int column() {
         return column;
@@ -104,7 +104,7 @@ public final class TextPos implements Serializable {
     }
 
     /**
-     * @return offset
+     * @return offset in 16-bit characters
      */
     public long offset() {
         return offset;
